requirements.txt
alembic==1.6.2
azure-core==1.14.0
azure-storage-blob==12.8.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
cryptography==3.4.7
dominate==2.6.0
Flask==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Migrate==3.0.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
msrest==0.6.21
oauthlib==3.1.0
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.15
urllib3==1.26.4
visitor==0.1.3
Werkzeug==2.0.1
